+++ cms/drupal/themes/build.php
@@ -12,7 +12,7 @@ namespace Deployer;
 
 task('cms:drupal:themes:build', static function (): void {
     foreach (get('themes') as $theme) {
-        within($theme, static function (): void {
+        within(get('release') . '/' . $theme, static function (): void {
             // @todo Detect theme package manager & task runner. Run commands based on that data.
             run('npm install');
             run('gulp build');