+++ CHANGELOG.md
@@ -14,7 +14,7 @@ Updates should follow the [Keep a CHANGELOG](https://keepachangelog.com/) princi
   variable can be overridden if standard-fare Drupal configuration via settings files
   is preferred. See [unleashedtech/dotenv-drupal](https://github.com/unleashedtech/dotenv-drupal) for more.
 - `shared_dir_names`, `shared_file_names` & `writable_dir_names` now require arrays of
-  absolute paths. Each path will be parsed for placeholders before use.
+  paths relative to the release root. Each path will be parsed for placeholders before use.
 - The `cms:drupal:db:backup:create` task now supports creating backups of multi-site apps.
 
 ### Fixed